## Major rewrite with improved shrinking
 - ?LET shrinks correctly!
 - rename module triq_domain -> triq_dom
 - move  code from triq_simplify.erl into triq_dom.erl 
 - rename things "simplify" -> "shrink"
 - rename things "generate" -> "pick"
 - added a bunch of extra unit tests.

We may have lost some of the edocs in the transition; I'll work to reintroduce those.

## details

The key to this update is that the two function pick(Domain,Size) and shrink(Domain,Value) now return a *pair* of {Domain,Value}, so that we can represent choices made during generation (picking a value from a domain) or shrinking in the result.  This also means that we can control shrinking much better to manage the structure and the order of shrinking.

As an example

  triq_dom:pick(list(oneof([int(),boolean()])), 3) 

may yield 

  {
    #'@'{kind=#shrinkable_list{template=[boolean(),int()]}, empty_ok=true},
    [false, 2] 
  }

i.e., the output of pick contains the structure of the result "[false, 4]", so that we can control the shrinking to be the right sequence... such as 

  [false, 1], 
  [false, 0], 
  [false] or [0],  
  [].