Increased level of randomness
The Erlang random number generator does not have an high level of randomness because it is based on a default seed and the result is that each Triq run generates the same random values.
I have created a triq:generate_randomness/0 function which uses crypto to improve the randomness and it is invoked in triq:check/3 which (if I am not wrong) is the function that finally checks a property.

So Triq now generates random values with a different seed for each property.