Fix int/2 generator
The issue was that the maximum value picked for the generator was
the highest between SampleSize (which is potentially very big)
and Diff (which is the real maximum we don't want to cross).
Instead we now pick the lowest, ensuring that we won't generate
a value bigger than Diff.

Tests suggested by @krestenkrab.