Add preliminary support for triq:conjunction/1.
Though this appears to work, I have some concerns about whether the
information that makes it out of this is going to be useful.  If you
have multiple clauses in the conjunction statement, you don't
necessarily know what part failed.  I'm also a little worried that we
are bypassing some of the property-test setup logic in check/1.  I don't
know whether that will cause problems down the road.