First implementation of large integer support.
I am not 100% sure that shrinking by orders of magnitude is the right
choice here.  It means that there will be large gaps in what is being
tested as part of the quickcheck process when large integers are used.
That being said, the naive methodology of simply subtracting one is
clearly unworkable.