Introduce and use compat random number module
In order for Triq to work with previous and current OTP releases, we
generate the triq_rnd module on first load of triq.beam. triq_rnd is
generated dynamically to make sure that we use the right random module
(either rand if available or else random). It only wraps the common
subset of the API, but that's sufficient for Triq to function.

We also have to conditionally define a macro for using
crypto:strong_rand_bytes/1 if available or else fall back to
crypto:rand_bytes/1. Otherwise, warnings_as_errors would break the
build.